refactor: Filter 컴포넌트 명 FilterTrigger로 변경
