+++ frontend/src/pages/list/FilterContainer/FilterTrigger/FilterBox/FilterBox.styled.ts
